+++ RELEASE-NOTES.md
@@ -26,6 +26,7 @@
 * #490 ECharts: BarSeriesOpts.barWidth
 * #493 IntReverseSequenceSeries - decreasing sequential IntSeries
 * #494 Upgrade to Parquet 1.15.2
+* #495 Force column name deduplication in SqlLoader
 
 ## Release 2.0.0-M2